refactor(sicp): move Chatbot from JS layout to Python layout
