+++ src/new_routes/sicpjs/_layout.tsx
@@ -5,9 +5,6 @@ import classNames from 'classnames';
 import { useEffect, useRef } from 'react';
 import { Outlet, useLocation, useNavigate, useParams } from 'react-router';
 import { readLocalStorage } from 'src/commons/hooks/useLocalStorageState';
-import Constants from 'src/commons/utils/Constants';
-import { useSession } from 'src/commons/utils/Hooks';
-import type { SicpSection } from 'src/features/sicp/chatCompletion/chatCompletion';
 import { CodeSnippetProvider } from 'src/features/sicp/CodeSnippetProvider';
 import { ParseJsonError } from 'src/features/sicp/parser/ParseJson';
 import { scrollRefIntoView } from 'src/features/sicp/utils/SicpUtils';
@@ -19,42 +16,16 @@ import {
 
 import SicpErrorBoundary from '../../features/sicp/errors/SicpErrorBoundary';
 import getSicpError, { SicpErrorType } from '../../features/sicp/errors/SicpErrors';
-import Chatbot from '../../pages/sicp/subcomponents/chatbot/Chatbot';
 
 const loadingComponent = <NonIdealState title="Loading Content" icon={<Spinner />} />;
 
-const getText = () => {
-  const divs = document.querySelectorAll('p.sicp-text');
-  let visibleParagraphs = '';
-
-  divs.forEach(div => {
-    const rect = div.getBoundingClientRect();
-
-    if (
-      rect.top <= window.innerHeight &&
-      rect.bottom >= 0 &&
-      rect.left <= window.innerWidth &&
-      rect.right >= 0
-    ) {
-      const text = div.textContent;
-      visibleParagraphs += text + '\n';
-    }
-  });
-
-  return visibleParagraphs;
-};
-
-function getSection() {
-  // To discard the '/sicpjs/'
-  return location.pathname.replace('/sicpjs/', '') as SicpSection;
-}
+// Chatbot utilities removed from JS layout
 
 function SicpLayout() {
   const { section } = useParams<{ section: string }>();
   const parentRef = useRef<HTMLDivElement>(null);
   const navigate = useNavigate();
   const location = useLocation();
-  const { isLoggedIn } = useSession();
 
   // Handle rerouting to the latest viewed section when clicking from the main
   // application navbar. Navigate replace logic is used to allow the user to
@@ -105,9 +76,7 @@ function SicpLayout() {
           )}
         </CodeSnippetProvider>
       </SicpErrorBoundary>
-      {isLoggedIn && Constants.featureFlags.enableSicpChatbot && (
-        <Chatbot getSection={getSection} getText={getText} />
-      )}
+      {/* Chatbot moved to Python layout */}
     </div>
   );
 }
