+++ src/new_routes/sicpy/_layout.tsx
@@ -5,13 +5,14 @@ import classNames from 'classnames';
 import { useEffect, useRef } from 'react';
 import { Outlet, useLocation, useNavigate, useParams } from 'react-router';
 import { readLocalStorage } from 'src/commons/hooks/useLocalStorageState';
+import { useSession } from 'src/commons/utils/Hooks';
 import { CodeSnippetProvider } from 'src/features/sicp/CodeSnippetProvider';
 import { ParseJsonError } from 'src/features/sicp/parser/ParseJson';
 import { scrollRefIntoView } from 'src/features/sicp/utils/SicpUtils';
 import {
-  SICPY_CACHE_KEY,
-  SICPY_INDEX,
-  useSicPySectionQuery,
+    SICPY_CACHE_KEY,
+    SICPY_INDEX,
+    useSicPySectionQuery,
 } from 'src/features/textbook/hooks/useTextbookSectionQuery';
 
 import SicpErrorBoundary from '../../features/sicp/errors/SicpErrorBoundary';
@@ -19,11 +20,39 @@ import getSicpError, { SicpErrorType } from '../../features/sicp/errors/SicpErro
 
 const loadingComponent = <NonIdealState title="Loading Content" icon={<Spinner />} />;
 
+const getText = () => {
+  const divs = document.querySelectorAll('p.sicp-text');
+  let visibleParagraphs = '';
+
+  divs.forEach(div => {
+    const rect = div.getBoundingClientRect();
+
+    if (
+      rect.top <= window.innerHeight &&
+      rect.bottom >= 0 &&
+      rect.left <= window.innerWidth &&
+      rect.right >= 0
+    ) {
+      const text = div.textContent;
+      visibleParagraphs += text + '\n';
+    }
+  });
+
+  return visibleParagraphs;
+};
+
 function SicPyLayout() {
   const { section } = useParams<{ section: string }>();
   const parentRef = useRef<HTMLDivElement>(null);
   const navigate = useNavigate();
   const location = useLocation();
+
+  const { isLoggedIn } = useSession();
+
+  function getSection() {
+    // To discard the '/sicpy/'
+    return location.pathname.replace('/sicpy/', '');
+  }
 
   // Handle rerouting to the latest viewed section when clicking from the main
   // application navbar. Navigate replace logic is used to allow the user to
@@ -73,6 +102,9 @@ function SicPyLayout() {
           )}
         </CodeSnippetProvider>
       </SicpErrorBoundary>
+        {isLoggedIn && Constants.featureFlags.enableSicpChatbot && (
+          <Chatbot getSection={getSection} getText={getText} />
+        )}
     </div>
   );
 }
